Remove runtime references to Mix in application
Mix is not available in production releases, so these references to Mix
prevented deploy to fly.io

The Dotenv.load was moved to config/runtime, as that seems like a more
appropriate spot -- it _is_ setting up configuration, and it's happening
at runtime.  :)

References to Mix in the router should be okay, as the router is
evaluated during compilation, not runtime.